requirements.txt
ansible >= 10.2.0
click >= 8.1.0
click-default-group >= 1.2.0
matplotlib >= 3.9.0
osquery >= 3.1.0
plotext >= 5.3.0
pre-commit >= 4.0
polars-lts-cpu >= 1.22.0
psutil >= 5.9.0
pyright == 1.1.379
ruff >= 0.6.4
scipy >= 1.10.0
shfmt-py >= 3.7.0
typing_extensions >= 4.11.0
pymongo >= 3.5.1
pytimeparse >= 1.1.8
